Ignore errors during autoload
Because of the usage of the prepend behaviour with spl_autoload_register the FF autoloader causes
a heap of warnings written to the Magento system log although the classes will be found by other
autoloaders. Ignoring errors should not be problematic beause failing to load a class will likely
fail horribly in a different place either way and we could not find a place where another autoloader
did not load the required classes.